--- src/app/utils/coordinate.jsx
@@ -1,22 +1,0 @@
-export const validCoord = {
-    0: [1, 10, 11],
-    10: [0, 1, 11, 20, 21],
-    20: [10, 11, 21, 30, 31],
-    30: [20, 21, 31, 40, 41],
-    40: [30, 31, 41, 50, 51],
-    50: [40, 41, 51, 60, 61],
-    60: [50, 51, 61, 70, 71],
-    70: [60, 61, 71, 80, 81],
-    80: [70, 71, 81, 90, 91],
-    90: [80, 81, 91],
-    9: [8, 18, 19],
-    19: [8, 9, 18, 28, 29],
-    29: [18, 19, 28, 38, 39],
-    39: [28, 29, 38, 48, 49],
-    49: [38, 39, 48, 58, 59],
-    59: [48, 49, 58, 68, 69],
-    69: [58, 59, 68, 78, 79],
-    79: [68, 69, 78, 88, 89],
-    89: [78, 79, 88, 98, 99],
-    99: [88, 89, 98],
-};
